// src/api/generated/models/Supplier.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Supplier response model
 */
export type Supplier = {
    name: string;
    contact_name?: (string | null);
    email?: (string | null);
    phone?: (string | null);
    address?: (string | null);
    website?: (string | null);
    payment_terms?: (string | null);
    notes?: (string | null);
    is_active?: boolean;
    id: string;
    business_id: string;
    created_at: string;
    updated_at: string;
};